+++ components/subframe/components/CustomTreeView.tsx
@@ -49,7 +49,7 @@ const Folder = React.forwardRef<HTMLElement, FolderProps>(function Folder(
 
 interface ItemProps extends React.HTMLAttributes<HTMLDivElement> {
   label?: string
-  value?: string
+  value?: string | React.ReactNode
   className?: string
 }
 
@@ -60,19 +60,19 @@ const Item = React.forwardRef<HTMLElement, ItemProps>(function Item(
   return (
     <div
       className={SubframeCore.twClassNames(
-        "group/ebb17dce flex w-full cursor-pointer items-center gap-2 rounded-md px-2 py-2 hover:bg-transparent",
+        "group/ebb17dce w-full cursor-pointer items-center gap-2 rounded-md px-2 py-2 hover:bg-transparent",
         className
       )}
       ref={ref as any}
       {...otherProps}
     >
       {label ? (
-        <span className="line-clamp-1 shrink-0 grow basis-0 font-caption text-caption text-subtext-color group-hover/ebb17dce:text-default-font">
+        <span className="line-clamp-1 shrink-0 font-caption text-caption text-subtext-color group-hover/ebb17dce:text-default-font">
           {label}
         </span>
       ) : null}
       {value ? (
-        <span className="text-right font-caption text-caption text-subtext-color group-hover/ebb17dce:text-default-font">
+        <span className="text-right font-caption text-caption text-subtext-color group-hover/ebb17dce:text-default-font justify-end float-right">
           {value}
         </span>
       ) : null}
